# [draft] [stable28] perf(rich_workspace): only add property for parent